// src/types/Movie.ts
export interface Movie  {
    maPhim: number
    tenPhim: string
    biDanh: string
    trailer: string
    hinhAnh: string
    moTa: string
    maNhom: string
    ngayKhoiChieu: string
    danhGia: number
    hot: boolean
    dangChieu: boolean
    sapChieu: boolean
}


export interface MovieDetail {
    maPhim: number;
    tenPhim: string;
    biDanh: string;
    trailer: string;
    hinhAnh: string;
    moTa: string;
    maNhom: string;
    hot: boolean;
    dangChieu: boolean;
    sapChieu: boolean;
    ngayKhoiChieu: string;
    danhGia: number;
}